Clean

+++ sources/light/Ray.hpp
@@ -12,7 +12,7 @@ public:
 
 	Ray(const Point& origin, const Vector& direction): origin(origin), direction(direction) {}
 
-	Point at(double t) const
+	Point at(float t) const
 	{
 		return origin + t * direction;
 	}
